% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeBasisFunctions.R
\name{analyzeBasisFunctions}
\alias{analyzeBasisFunctions}
\title{Enable the user to estimate the optimal number of basis functions.}
\usage{
analyzeBasisFunctions(
  sp.obj = NULL,
  naive.hvgs = NULL,
  lscale.estimator = "kmeans",
  n.basis.fns = 20L,
  kernel = "exp_quad",
  kernel.smoothness = 1.5,
  kernel.period = 100,
  facet.ncol = NULL,
  pt.size = 1,
  n.cores = 2L,
  random.seed = 312
)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} or \code{SpatialExperiment} containing spatial data. Defaults to NULL.}

\item{naive.hvgs}{A vector containing genes that have been classified as naive HVGs from which SVGs will be detected. Defaults to NULL.}

\item{lscale.estimator}{A string specifying how the global length-scale should be estimated. Must be one of "kmeans" or "variogram". Defaults to "kmeans".}

\item{n.basis.fns}{An integer specifying the number of basis functions to be used when approximating the GP as a Hilbert space. Defaults to 20.}

\item{kernel}{A string specifying the covariance kernel to be used when fitting the GP. Must be one of "exp_quad", "matern", or "periodic". Defaults to "exp_quad".}

\item{kernel.smoothness}{A double specifying the smoothness parameter \eqn{\nu} used when computing the Matern kernel. Must be one of 0.5, 1.5, or 2.5. Using 0.5 corresponds to the exponential kernel. Defaults to 1.5.}

\item{kernel.period}{An integer specifying the period parameter \eqn{p} used when computing the periodic kernel. Defaults to 100.}

\item{facet.ncol}{An integer specifying the number of columns in faceted plots generated with \code{\link[ggplot2]{facet_wrap}}. Defaults to NULL.}

\item{pt.size}{A double specifying the size of the points in scatterplots generated with \code{\link[ggplot2]{geom_point}}. Defaults to 1.}

\item{n.cores}{An integer specifying the number of threads used when estimating the length-scale via the variogram method. Defaults to 2.}

\item{random.seed}{A double specifying the random seed to be set prior to running \emph{k}-means. Defaults to 312.}
}
\value{
A named list containing plots and summary statistics concerning the structure & quality of the generated basis functions.
}
\description{
Given a \code{Seurat} or \code{SpatialExperiment} object, generate a variety of statistics & plots that will help you to choose the optimal number of basis functions \emph{k} for usage in \code{\link{findSpatiallyVariableFeaturesBayes}}.
}
\details{
\itemize{
\item This function replicates the logic & methods used within \code{\link{findSpatiallyVariableFeaturesBayes}} to generate the basis function matrices \eqn{\Phi} (non-orthonormal) and \eqn{\tilde{\Phi}} (orthonormal). 
\item The first three elements of the returned list are visualizations of the structure of the basis functions: spatial scatterplots colored by \emph{k}-means cluster ID, basis function values, and basis function magnitudes, respectively. 
\item The fourth element of the returned list is a visualization of basis function eigenvalue decay. The x-axis is simply the index of the basis functions from \eqn{j = 1, \dots, k}, while the y-axis is the log of the eigenvalues \eqn{\lambda_j} of \eqn{\Phi\Phi^\top}. Swift eigenvalue decay indicates that the effective dimension of the spatial process is small.  
\item The next three elements of the returned list are \eqn{\Phi} (the non-orthonormal basis function matrix), \eqn{\tilde{\Phi}} (the QR-decomposed basis function matrix), and the output generated by running \code{\link[stats]{kmeans}} on the scaled spatial coordinates. 
\item The returned list also include an estimate of the Frobenius error of the low-rank approximation of the spatial covariance matrix \eqn{\hat{K} = \tilde{\Phi}^\top \tilde{\Phi}} when compared to the full spatial covariance matrix \eqn{K}. The Frobenius error is computed as: \eqn{\text{err} = \frac{||K - \hat{K}||_F}{||K||_F}}. 
\item The final item in the returned list is a sub-list containing all the relevant parameters used in the generation of the basis functions. This can be useful when running large experiments with many repetitions of the kernel selection procedure.
}
}
\examples{
data(seu_brain)
seu_brain <- Seurat::NormalizeData(seu_brain, verbose = FALSE)
naive_hvgs <- getNaiveHVGs(seu_brain, n.hvg = 500L)
analyzeBasisFunctions(seu_brain, naive.hvgs = naive_hvgs)
}
\seealso{
\code{\link{plotBasisFunctions}}

\code{\link{findSpatiallyVariableFeaturesBayes}}
}
\author{
Jack R. Leary
}
