% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyHVGs.R
\name{classifyHVGs}
\alias{classifyHVGs}
\title{Classify genes as highly variable.}
\usage{
classifyHVGs(
  sc.obj = NULL,
  selection.variable = "dispersion",
  selection.method = "rank",
  n.HVG = 2000L,
  quantile.HVG = 0.75,
  cutoff = 3
)
}
\arguments{
\item{sc.obj}{An object of class \code{Seurat} or \code{SingleCellExperiment}. Defaults to NULL.}

\item{selection.variable}{A string specifying the random variable used to select HVGs. Must be one of "dispersion" or "sigma2". Defaults to "dispersion".}

\item{selection.method}{A string specifying what method should be used to classify genes as HVGs. Must be one of "rank", "quantile", or "cutoff". Defaults to "rank".}

\item{n.HVG}{An integer specifying the number of HVGs to select (if using rank-based selection). Defaults to 2000.}

\item{quantile.HVG}{A double specifying the quantile cutoff used to classify HVGs (if using quantile-based selection). Defaults to 0.75.}

\item{cutoff}{A double specifying the cutoff value for dispersion or variance (depending on how \code{selection.variable} is defined) used to classify HVGs (if using cutoff-based selection). Defaults to 3.}
}
\value{
Depending on the input, either an object of class \code{Seurat} or \code{SingleCellExperiment} with HVG metadata added.
}
\description{
After estimating per-gene statistics, classify genes as highly variable or not in a variety of ways.
}
\seealso{
\code{\link{findVariableFeaturesBayes}}

\code{\link[SeuratObject]{HVFInfo}}
}
\author{
Jack R. Leary
}
