% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySVGs.R
\name{classifySVGs}
\alias{classifySVGs}
\title{Classify genes as spatially variable.}
\usage{
classifySVGs(
  sp.obj = NULL,
  selection.method = "rank",
  n.SVG = 1000L,
  quantile.SVG = 0.75,
  cutoff = 0.05
)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} containing spatial data. Defaults to NULL.}

\item{selection.method}{A string specifying what method should be used to classify genes as SVGs. Must be one of "rank", "quantile", or "cutoff". Defaults to "rank".}

\item{n.SVG}{An integer specifying the number of SVGs to select (if using rank-based selection). Defaults to 1000.}

\item{quantile.SVG}{A double specifying the quantile cutoff used to classify SVGs (if using quantile-based selection). Defaults to 0.75.}

\item{cutoff}{A double specifying the cutoff value for spatial variation (depending on how \code{selection.variable} is defined) used to classify SVGs (if using cutoff-based selection). Defaults to 0.05.}
}
\value{
An object of class \code{Seurat} with SVG metadata added.
}
\description{
After estimating per-gene spatial variation statistics, classify genes as spatially variable or not in a variety of ways.
}
\seealso{
\code{\link{findSpatiallyVariableFeaturesBayes}}

\code{\link[SeuratObject]{SVFInfo}}
}
\author{
Jack R. Leary
}
