% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToSpatialExperiment.R
\name{convertToSpatialExperiment}
\alias{convertToSpatialExperiment}
\title{Convert from \code{Seurat} to \code{SpatialExperiment}.}
\usage{
convertToSpatialExperiment(
  seu.obj = NULL,
  sample.id = NULL,
  scale.coords = FALSE
)
}
\arguments{
\item{seu.obj}{An object of class \code{Seurat}. Defaults to NULL.}

\item{sample.id}{A string specifying the sample ID corresponding to the image saved in \code{seu.obj}.}

\item{scale.coords}{A Boolean specifying whether the spatial coordinates matrix should be scaled. Defaults to FALSE.}
}
\value{
An object of class \code{SpatialExperiment}.
}
\description{
This function converts a \code{Seurat} object to a \code{SpatialExperiment} object while retaining all assays, metadata, and images.
}
\examples{
data(seu_brain)
spe_brain <- convertToSpatialExperiment(seu_brain, sample.id = "anterior1")
}
\seealso{
\code{\link[Seurat]{as.SingleCellExperiment}}

\code{\link[SpatialExperiment]{SpatialExperiment}}
}
\author{
Jack R. Leary
}
