% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establishAmplitudeThreshold.R
\name{establishAmplitudeThreshold}
\alias{establishAmplitudeThreshold}
\title{Identify spatially variable genes in a Bayesian manner.}
\usage{
establishAmplitudeThreshold(
  sp.obj = NULL,
  sampled.genes = NULL,
  kernel = "exp_quad",
  kernel.smoothness = 1.5,
  kernel.period = 100L,
  n.basis.fns = 20L,
  algorithm = "meanfield",
  mle.init = TRUE,
  gene.depth.adjust = TRUE,
  n.cores = 2L,
  random.seed = 312,
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} containing spatial data. Defaults to NULL.}

\item{sampled.genes}{A vector containing a subsample of the rownames of \code{sp.obj}. Defaults to NULL.}

\item{kernel}{A string specifying the covariance kernel to be used when fitting the GP. Must be one of "exp_quad", "matern", or "periodic". Defaults to "exp_quad".}

\item{kernel.smoothness}{A double specifying the smoothness parameter \eqn{\nu} used when computing the Matérn kernel. Must be one of 0.5, 1.5, or 2.5. Using 0.5 corresponds to the exponential kernel. Defaults to 1.5.}

\item{kernel.period}{An integer specifying the period parameter \eqn{p} used when computing the periodic kernel. Defaults to 100.}

\item{n.basis.fns}{An integer specifying the number of basis functions to be used when approximating the GP as a Hilbert space. Defaults to 20.}

\item{algorithm}{A string specifying the variational inference (VI) approximation algorithm to be used. Must be one of "meanfield", "fullrank", or "pathfinder". Defaults to "meanfield".}

\item{mle.init}{A Boolean specifying whether the the VI algorithm should be initialized using the MLE for each parameter. In general, this isn't strictly necessary but can help if the VI algorithm struggles to converge when provided with the default initialization (zero). Cannot be used when the Pathfinder algorithm is specified. Defaults to TRUE.}

\item{gene.depth.adjust}{A Boolean specifying whether the model should include a fixed effect term for total gene expression. Defaults to TRUE.}

\item{n.cores}{An integer specifying the number of threads used in compiling and fitting the model. Defaults to 2.}

\item{random.seed}{A double specifying the random seed to be used when fitting and sampling from the model. Defaults to 312.}

\item{verbose}{A Boolean specifying whether or not verbose model output should be printed to the console. Defaults to TRUE.}
}
\value{
A \code{data.frame} containing amplitude-related statistics for each gene.
}
\description{
This function fits the same Gaussian process model as \code{\link{findSpatiallyVariableFeaturesBayes}}, but does so for a subset of genes on shuffled spatial coordinates.
}
\details{
\itemize{
\item This function is meant to be run in tandem with \code{\link{findSpatiallyVariableFeaturesBayes}} in order to determine an accurate threshold for the gene-specific amplitude parameter \eqn{\tau_g}.
}
}
\examples{
data(seu_brain)
seu_brain <- Seurat::SCTransform(seu_brain,
                                 assay = "Spatial",
                                 variable.features.n = 3000L,
                                 vst.flavor = "v2",
                                 return.only.var.genes = FALSE,
                                 seed.use = 312,
                                 verbose = FALSE)
set.seed(312)
sampled_genes <- sample(VariableFeatures(seu_brain), size = 50L)
amplitude_summary <- establishAmplitudeThreshold(seu_brain, sampled.genes = sampled_genes)
}
\seealso{
\code{\link{findSpatiallyVariableFeaturesBayes}}
}
\author{
Jack R. Leary
}
