% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expQuadKernel.R
\name{expQuadKernel}
\alias{expQuadKernel}
\title{Compute the exponentiated quadratic kernel.}
\usage{
expQuadKernel(d = NULL, length.scale = NULL)
}
\arguments{
\item{d}{A numeric vector of distances. Defaults to NULL.}

\item{length.scale}{A numeric specifying the length-scale \eqn{\ell} for the approximate GP. Defaults to NULL.}
}
\value{
A vector of numeric values.
}
\description{
This function computes the exponentiated quadratic kernel for a given vector of distances and a global length-scale.
}
\seealso{
\code{\link{maternKernel}}
}
\author{
Jack R. Leary
}
