% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNaiveHVGs.R
\name{getNaiveHVGs}
\alias{getNaiveHVGs}
\title{Compute a set of naive HVGs.}
\usage{
getNaiveHVGs(sp.obj = NULL, n.hvg = 3000L)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} or \code{SpatialExperiment} containing spatial data. The function assumes that the raw counts have already been depth- and log1p-normalized. Defaults to NULL.}

\item{n.hvg}{An integer specifying the number of naive HVGs to be returned. Defaults to 3000.}
}
\value{
A character vector of length \code{n.hvg} specifying the set of naive HVGs.
}
\description{
This function, given an object of class \code{Seurat} or \code{SpatialExperiment}, computes a set of naive HVGs to be used as candidates for SVG modeling.
}
\examples{
data(seu_brain)
seu_brain <- Seurat::NormalizeData(seu_brain, verbose = FALSE)
naive_hvgs <- getNaiveHVGs(seu_brain)
}
\seealso{
\code{\link[Seurat]{FindVariableFeatures}}

\code{\link[scrapper]{chooseRnaHvgs.se}}
}
\author{
Jack R. Leary
}
