% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternKernel.R
\name{maternKernel}
\alias{maternKernel}
\title{Compute the Matern-family kernel.}
\usage{
maternKernel(d = NULL, length.scale = NULL, nu = NULL, sigma = 1)
}
\arguments{
\item{d}{A numeric vector of distances. Defaults to NULL.}

\item{length.scale}{A numeric specifying the length-scale \eqn{\ell} for the approximate GP. Defaults to NULL.}

\item{nu}{A numeric specifying the smoothness parameter \eqn{\nu} of the Matern kernel. Defaults to NULL.}

\item{sigma}{A numeric specifying the variance of the Matern kernel. Defaults to 1.}
}
\value{
A vector of numeric values.
}
\description{
This function computes the Matern kernel for a given vector of distances and a global length-scale.
}
\seealso{
\code{\link{expQuadKernel}}
}
\author{
Jack R. Leary
}
