% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicKernel.R
\name{periodicKernel}
\alias{periodicKernel}
\title{Compute the periodic kernel.}
\usage{
periodicKernel(d = NULL, length.scale = NULL, period = 100L)
}
\arguments{
\item{d}{A numeric vector of distances. Defaults to NULL.}

\item{length.scale}{A numeric specifying the length-scale \eqn{\ell} for the approximate GP. Defaults to NULL.}

\item{period}{An integer specifying the period of the kernel. Defaults to 100.}
}
\value{
A vector of numeric values.
}
\description{
This function computes the periodic kernel for a given vector of distances and a global length-scale.
}
\seealso{
\code{\link{maternKernel}}

\code{\link{expQuadKernel}}
}
\author{
Jack R. Leary
}
