% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHVGs.R
\name{plotHVGs}
\alias{plotHVGs}
\title{Plot the results of HVG identification.}
\usage{
plotHVGs(
  sc.obj = NULL,
  pt.size = 1,
  pt.alpha = 0.6,
  add.smooth = TRUE,
  n.genes.label = 10L,
  label.text.size = 3
)
}
\arguments{
\item{sc.obj}{An object of class \code{Seurat} or \code{SingleCellExperiment} upon which \code{\link{findVariableFeaturesBayes}} and \code{\link{classifyHVGs}} have been run. Defaults to NULL.}

\item{pt.size}{A double specifying the size of the points on the plot. Defaults to 1.}

\item{pt.alpha}{A double specifying the opacity of the points on the plot. Defaults to 0.6.}

\item{add.smooth}{A Boolean specifying whether a GAM smoother should be overlaid in order to show the overall relationship between the (log) mean and (log) dispersion of gene expression. Defaults to TRUE.}

\item{n.genes.label}{An integer specifying the number of top HVGs to label on the plot using \code{\link[ggrepel]{geom_label_repel}}. If equal to 0, no genes will be labelled. Defaults to 10.}

\item{label.text.size}{A double specifying the size of the text for each top HVG's label. Defaults to 3.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
This function plots the (log) mean versus the (log) dispersion of gene expression in order to visualize the relationship between the two statistics for HVGs and non-HVGs.
}
\details{
\itemize{
\item When multiple subjects are present (this is detected automatically by internal logic), the resulting plot will be faceted by subject and the top \code{n.genes.label} per-subject will be labelled. 
}
}
\examples{
data(seu_pbmc)
seu_pbmc <- findVariableFeaturesBayes(seu_pbmc,
                                      n.cells.subsample = 500L,
                                      algorithm = "meanfield",
                                      n.cores.per.chain = 1L,
                                      save.model = TRUE) \%>\%
            classifyHVGs(n.HVG = 500L)
plotHVGs(seu_pbmc)
}
\seealso{
\code{\link{findVariableFeaturesBayes}}

\code{\link{classifyHVGs}}

\code{\link[Seurat]{VariableFeaturePlot}}
}
\author{
Jack R. Leary
}
