% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialAttributes.R
\name{plotSpatialAttributes}
\alias{plotSpatialAttributes}
\title{Generate a plot of annotated spots.}
\usage{
plotSpatialAttributes(
  sp.obj = NULL,
  attribute.plot = NULL,
  pt.size = 2,
  color.palette = NULL
)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} or \code{SpatialExperiment} containing spatial data. Defaults to NULL.}

\item{attribute.plot}{A character specifying the categorical feature to be used to color the spots. Defaults to NULL.}

\item{pt.size}{A double specifying the size of the points to be plotted. Defaults to 2.}

\item{color.palette}{A vector containing colors that are passed to \code{\link[ggplot2]{scale_color_manual}} and thus used to color the spots. Defaults to NULL.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
Given a \code{Seurat} or \code{SpatialExperiment} object, plot the spatial location of each spot colored by a discrete attribute such as cluster ID, annotated domain, etc.
}
\examples{
data(seu_brain)
plotSpatialAttributes(seu_brain, attribute.plot = "region")
}
\seealso{
\code{\link[Seurat]{SpatialDimPlot}}

\code{\link[ggspavis]{plotCoords}}

\code{\link{plotSpatialExpression}}
}
\author{
Jack R. Leary
}
