% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialExpression.R
\name{plotSpatialExpression}
\alias{plotSpatialExpression}
\title{Generate a plot of annotated spots.}
\usage{
plotSpatialExpression(
  sp.obj = NULL,
  gene.plot = NULL,
  use.norm = TRUE,
  scale.expression = NULL,
  pt.size = 2,
  color.palette = NULL
)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} or \code{SpatialExperiment} containing spatial data. Defaults to NULL.}

\item{gene.plot}{A character or character vector specifying the gene(s) whose expression will be used to color the spots. Defaults to NULL.}

\item{use.norm}{A Boolean specifying whether the raw or normalized counts should be plotted. Defaults to TRUE.}

\item{scale.expression}{A Boolean specifying whether expression should be scaled prior to plotting. Mostly relevant for when \code{gene.plot} contains more than one value. Defaults to TRUE when \code{length(gene.plot) > 1} and FALSE otherwise.}

\item{pt.size}{A double specifying the size of the points to be plotted. Defaults to 2.}

\item{color.palette}{A vector containing colors that are passed to \code{\link[ggplot2]{scale_color_gradientn}} and thus used to color the spots. Defaults to NULL.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
Given a \code{Seurat} or \code{SpatialExperiment} object, plot the spatial location of each spot colored by the expression of a gene of interest.
}
\examples{
data(seu_brain)
plotSpatialExpression(seu_brain,
                      gene.plot = "Nrgn",
                      use.norm = FALSE)
}
\seealso{
\code{\link[Seurat]{SpatialFeaturePlot}}
}
\author{
Jack R. Leary
}
