% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreSpatialModules.R
\name{scoreSpatialModules}
\alias{scoreSpatialModules}
\title{Estimate module score for spatial gene sets.}
\usage{
scoreSpatialModules(sp.obj = NULL, svg.clusters = NULL, n.cores = 2L)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} or \code{SpatialExperiment} upon which \code{\link{findSpatiallyVariableFeaturesBayes}} and \code{\link{classifySVGs}} have been run. Defaults to NULL.}

\item{svg.clusters}{The results from \code{\link{clusterSVGsBayes}}. Defaults to NULL.}

\item{n.cores}{An integer specifying the number of cores used when running \code{UCell}. Defaults to 2.}
}
\value{
Depending on the input, either an object of class \code{Seurat} or \code{SpatialExperiment} with per-cluster module scores added to the appropriate metadata slot. Module scores are named as e.g., \code{"svg_cluster_1_UCell"}.
}
\description{
Given a \code{Seurat} or \code{SpatialExperiment} and the results from \code{\link{clusterSVGsBayes}}, this function uses \code{UCell} to estimate gene set scores for each cluster of SVGs.
}
\examples{
data(seu_brain)
seu_brain <- Seurat::SCTransform(seu_brain,
                                 assay = "Spatial",
                                 variable.features.n = 3000L,
                                 vst.flavor = "v2",
                                 return.only.var.genes = FALSE,
                                 seed.use = 312,
                                 verbose = FALSE)
seu_brain <- findSpatiallyVariableFeaturesBayes(seu_brain,
                                                naive.hvgs = Seurat::VariableFeatures(seu_brain),
                                                kernel = "matern",
                                                kernel.smoothness = 1.5,
                                                algorithm = "meanfield",
                                                n.cores = 1L,
                                                save.model = TRUE) \%>\%
             classifySVGs(n.SVG = 1000L)
svg_clusters <- clusterSVGsBayes(seu_brain,
                                 svgs = Seurat::VariableFeatures(seu_brain),
                                 n.clust = 3L,
                                 n.cores = 1L)
seu_brain <- scoreSpatialModules(seu_brain, 
                                 svg.clusters = svg_clusters, 
                                 n.cores = 1L)
}
\author{
Jack R. Leary
}
