% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seu_brain}
\alias{seu_brain}
\title{A \code{Seurat} object containing spatial data.}
\format{
An object of class \code{Seurat}.
}
\source{
\url{https://github.com/satijalab/seurat-data}
}
\usage{
data(seu_brain)
}
\description{
This file contains the anterior mouse brain data from 10X Genomics.
}
\keyword{datasets}
