% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seu_pbmc}
\alias{seu_pbmc}
\title{A \code{Seurat} object containing scRNA-seq data.}
\format{
An object of class \code{Seurat}.
}
\source{
\url{https://github.com/satijalab/seurat-data}
}
\usage{
data(seu_pbmc)
}
\description{
This file contains the pbmc3k data from 10X Genomics.
}
\keyword{datasets}
