% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialKmeans.R
\name{spatialKmeans}
\alias{spatialKmeans}
\title{Run \emph{k}-means clustering on spatial coordinates with special conditions.}
\usage{
spatialKmeans(
  coord.mtx = NULL,
  n.centers = 20L,
  iter.max = 100L,
  n.start = 10L
)
}
\arguments{
\item{iter.max}{An integer specifying the maximum number of algorithm iterations. Defaults to 100.}

\item{n.start}{An integer passed to the `nstart` argument of \code{\link[stats]{kmeans}}. Defaults to 10.}

\item{coord_mtx}{A matrix of spatial coordinates. Defaults to NULL.}
}
\value{
An object of class \code{kmeans}.
}
\description{
This helper function runs \emph{k}-means clustering on the spatial coordinates matrix, with added functionalities that make convergence more likely.
}
\seealso{
\code{\link[stats]{kmeans}}
}
\author{
Jack R. Leary
}
