% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bayesVG.R
\name{theme_bayesVG}
\alias{theme_bayesVG}
\title{A \code{ggplot2} theme for \code{bayesVG}.}
\usage{
theme_bayesVG(
  base.size = 12,
  base.lwd = 0.75,
  base.family = "sans",
  umap = FALSE,
  spatial = FALSE
)
}
\arguments{
\item{base.size}{The base font size. Defaults to 12.}

\item{base.lwd}{The base linewidth. Defaults to 0.75.}

\item{base.family}{The font family to be used throughout. Defaults to "sans".}

\item{umap}{(Optional) If set to TRUE, removes axis text and ticks for a cleaner look. Defaults to FALSE.}

\item{spatial}{(Optional) If set to true, provides a box around the spatial representation of the data and cleans up the axes. Defaults to FALSE.}
}
\value{
A \code{ggplot2} theme.
}
\description{
A publication-ready theme for creating gene expression scatterplots, embedding plots, etc.
}
\author{
Jack R. Leary
}
