% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTempJSON.R
\name{writeTempJSON}
\alias{writeTempJSON}
\title{Write extremely large datasets to JSON for ingestion by Stan.}
\usage{
writeTempJSON(data.list = NULL, temp.json = NULL, chunk.size = 1e+06)
}
\arguments{
\item{data.list}{A named list containing the data to be modeled. Defaults to NULL.}

\item{temp.json}{The output of \code{tempfile(fileext = ".json")}. Defaults to NULL.}

\item{chunk.size}{A double specifying the length above which each element of \code{data.list} will be streamed to the temporary JSON file in chunks. Defaults to one million (\code{1e6}).}
}
\value{
\code{invisible(NULL)}.
}
\description{
In order to bypass R's vector length limits when writing to disk, this function uses chunking to write very large datasets to a temporary JSON file that can then be accessed by \code{CmdStan}.
}
\author{
Jack R. Leary
}
