% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSVGsBayes.R
\name{clusterSVGsBayes}
\alias{clusterSVGsBayes}
\title{Cluster SVGs into modules.}
\usage{
clusterSVGsBayes(sp.obj = NULL, svgs = NULL, n.clust = 5L)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} or \code{SpatialExperiment} upon which \code{\link{findSpatiallyVariableFeaturesBayes}} and \code{\link{classifySVGs}} have been run. Defaults to NULL.}

\item{svgs}{A character vector containing the identified SVGs. Defaults to NULL.}

\item{n.clust}{An integer specifying the number of clusters to fit to the data. Defaults to 5.}
}
\value{
A \code{data.frame} containing the per-SVG soft cluster assignments.
}
\description{
This downstream analysis function clusters identified SVGs into modules via an approximate Bayesian soft-clustering approach.
}
\author{
Jack R. Leary
}
