% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNaiveGeneStatistics.R
\name{computeNaiveGeneStatistics}
\alias{computeNaiveGeneStatistics}
\title{Compute the naive mean, variance, and dispersion of gene expression.}
\usage{
computeNaiveGeneStatistics(obj = NULL, use.norm = FALSE)
}
\arguments{
\item{obj}{An object of class \code{Seurat}, \code{SingleCellExperiment}, or \code{SpatialExperiment}. Defaults to NULL.}

\item{use.norm}{A Boolean specifying whether the normalized counts should be used instead of raw counts. Defaults to FALSE.}
}
\value{
A \code{data.frame} containing the per-gene naive mean, variance, and dispersion.
}
\description{
Useful as a comparison to the results generated by our Bayesian models, this function computes the naive mean, variance, and dispersion of gene expression.
}
\examples{
data(seu_pbmc)
naive_gene_stats <- computeNaiveGeneStatistics(seu_pbmc)
}
\author{
Jack R. Leary
}
