% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpatiallyVariableFeaturesBayes.R
\name{findSpatiallyVariableFeaturesBayes}
\alias{findSpatiallyVariableFeaturesBayes}
\title{Identify spatially variable genes in a Bayesian manner.}
\usage{
findSpatiallyVariableFeaturesBayes(
  sp.obj = NULL,
  n.iter = 3000L,
  kernel = "exp_quad",
  kernel.smoothness = 1.5,
  n.basis.fns = 20L,
  algorithm = "meanfield",
  n.draws = 1000L,
  opencl.params = NULL,
  random.seed = 312,
  verbose = TRUE,
  save.model = FALSE
)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} containing spatial data. Defaults to NULL.}

\item{kernel}{A string specifying the covariance kernel to be used when fitting the GP. Must be one of "exp_quad" or "matern". Defaults to "exp_quad".}

\item{kernel.smoothness}{A double specifing the smoothness parameter \eqn{\nu} used when computing the Matern kernel. Must be one of 0.5, 1.5, or 2.5. Using 0.5 corresponds to the exponential kernel. Defaults to 1.5.}

\item{n.basis.fns}{(Optional) An integer specifying the number of basis functions to be used when approximating the GP as a Hilbert space. Defaults to 20.}

\item{algorithm}{(Optional) A string specifying the variational inference approximation algorithm to be used. Must be one of "meanfield", "fullrank", or "pathfinder". Defaults to "meanfield".}

\item{opencl.params}{(Optional) A two-element double vector specifying the platform and device IDs of the OpenCL GPU device. Most users should specify \code{c(0, 0)}. See \code{\link[brms]{opencl}} for more details. Defaults to NULL.}

\item{random.seed}{A double specifying the random seed to be used when fitting and sampling from the model. Defaults to 312.}

\item{verbose}{(Optional) A Boolean specifying whether or not verbose model output should be printed to the console. Defaults to TRUE.}

\item{save.model}{(Optional) A Boolean specifying whether or not the fitted model generated by \code{\link[cmdstanr]{cmdstan_model}} should be saved to the unstructured metadata of \code{spatial.obj}. Defaults to FALSE.}

\item{iter.per.chain}{An integer specifying the number of iterations per chain. Defaults to 3000.}
}
\description{
This function implements SVG estimation using Bayesian variational inference to approximate the posterior distribution of each gene's expression while accounting for spatial trends.
}
\details{
\itemize{
\item This function utilizes an approximate multivariate hierarchical Gaussian process (GP) to model spatial variation in gene expression.
\item While we have implemented GPU acceleration via OpenCL through the argument \code{opencl.params}, OpenCL acceleration is not supported on every machine. For example, Apple M-series chips do not support double-precision floating-points, which are necessary for Stan to compile. For more information, see \href{https://discourse.mc-stan.org/t/gpus-on-mac-osx-apple-m1/23375/5}{this Stan forums thread}. In order to correctly specify the OpenCL platform and device IDs, use the \code{clinfo} command line utility.
\item The user can specify which variational inference (VI) to use to fit the model via the argument \code{algorithm}. For further details, see \href{https://www.jmlr.org/papers/volume18/16-107/16-107.pdf}{this paper} comparing the meanfield and fullrank algorithms, and \href{https://doi.org/10.48550/arXiv.2108.03782}{this preprint} that introduced the Pathfinder algorithm. For a primer on automatic differentiation variational inference (ADVI), see \href{https://doi.org/10.48550/arXiv.1506.03431}{this preprint}. Lastly, \href{https://discourse.mc-stan.org/t/issues-with-differences-between-mcmc-and-pathfinder-results-how-to-make-pathfinder-or-something-else-more-accurate/35992}{this Stan forums thread} lays out some pratical differences between the algorithms.
\item If \code{save.model} is set to TRUE, the final model fit will be saved to the appropriate unstructured metadata slot of \code{sp.obj}. This allows the user to inspect the final fit and perform posterior predictive checks, but the model object takes up a lot of space. As such, it is recommended to remove it from \code{sp.obj} by setting the appropriate slot to NULL before saving it to disk.
}
}
\seealso{
\code{\link[Seurat]{FindSpatiallyVariableFeatures}}
}
\author{
Jack R. Leary
}
