% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette_cluster}
\alias{palette_cluster}
\title{A discrete color palette.}
\format{
A character vector of hex color codes.
}
\usage{
data(palette_cluster)
}
\description{
This color palette is useful for plotting cluster or domain IDs.
}
\keyword{datasets}
