% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette_heatmap}
\alias{palette_heatmap}
\title{A continuous color palette.}
\format{
A character vector of hex color codes.
}
\source{
\code{paletteer::paletteer_d("MetBrewer::Hiroshige", direction = -1)}
}
\usage{
data(palette_heatmap)
}
\description{
This color palette is useful for plotting heatmaps or gene expression / score values.
}
\keyword{datasets}
