% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBasisFunctions.R
\name{plotBasisFunctions}
\alias{plotBasisFunctions}
\title{Generate a plot of estimated basis functions.}
\usage{
plotBasisFunctions(
  sp.obj = NULL,
  plot.type = "basis_spatial",
  facet.ncol = NULL,
  pt.size = 1,
  color.palette = NULL
)
}
\arguments{
\item{sp.obj}{An object of class \code{Seurat} or \code{SpatialExperiment} containing spatial data. Defaults to NULL.}

\item{plot.type}{A character specifying which type of plot should be generated. Must be one of "clustering", "basis_spatial", "basis_abs_spatial", or "basis_orthog". Defaults to "basis_spatial".}

\item{facet.ncol}{An integer specifying the number of columns in the faceted plot generated by \code{\link[ggplot2]{facet_wrap}} when \code{plot.type} is "basis_spatial" or "basis_abs_spatial". Defaults to NULL.}

\item{pt.size}{A double specifying the size of the points to be plotted when \code{plot.type} is "clustering", "basis_spatial", or "basis_abs_spatial". Defaults to 1.}

\item{color.palette}{A vector containing colors that are passed to \code{\link[ggplot2]{scale_color_gradientn}}, \code{\link[ggplot2]{scale_color_manual}}, \code{\link[ggplot2]{scale_fill_gradientn}}, depending on the value of \code{plot.type}. Defaults to NULL.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
Given a \code{Seurat} or \code{SpatialExperiment} object, generate a variety of plots showing the structure of the estimated basis functions used to approximate the GP.
}
\examples{
data(seu_brain)
seu_brain <- Seurat::NormalizeData(seu_brain, verbose = FALSE)
naive_hvgs <- getNaiveHVGs(seu_brain, n.hvg = 500L)
seu_brain <- findSpatiallyVariableFeaturesBayes(seu_brain,
                                                naive.hvgs = naive_hvgs,
                                                kernel = "matern",
                                                kernel.smoothness = 1.5,
                                                algorithm = "meanfield",
                                                n.cores = 1L,
                                                save.model = TRUE, 
                                                save.basis.fns = TRUE) \%>\%
             classifySVGs(n.SVG = 100L)
plotBasisFunctions(seu_brain, plot.type = "basis_spatial")
}
\seealso{
\code{\link{plotSpatialModules}}

\code{\link{plotSpatialExpression}}

\code{\link{plotSpatialAttributes}}
}
\author{
Jack R. Leary
}
