% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTissueImage.R
\name{plotTissueImage}
\alias{plotTissueImage}
\title{Generate a plot of the tissue image given a \code{Seurat} object.}
\usage{
plotTissueImage(seu.obj = NULL, image.name = NULL)
}
\arguments{
\item{seu.obj}{An object of class \code{Seurat} containing spatial data. Defaults to NULL.}

\item{image.name}{A string specifying the name of the image to extract from \code{seu.obj}. If left unspecified, an attempt will be made to select the correct image automatically. Defaults to NULL.}
}
\value{
An object of class \code{ggplot2}.
}
\description{
Given a \code{Seurat} object, plot a high-definition visual of the tissue image without any annotations.
}
\examples{
data(seu_brain)
plotTissueImage(seu_brain)
}
\seealso{
\code{\link[Seurat]{GetImage}}

\code{\link[Seurat]{SpatialFeaturePlot}}
}
\author{
Jack R. Leary
}
